--- app/src/main/java/com/codesoom/assignment/HTTPMethod.java
@@ -1,13 +1,0 @@
-package com.codesoom.assignment;
-
-public enum HTTPMethod {
-    GET, POST, PUT, PATCH, DELETE;
-
-    public static HTTPMethod convert(String method) {
-        try {
-            return HTTPMethod.valueOf(method);
-        } catch (final IllegalArgumentException e) {
-            return null;
-        }
-    }
-}
