+++ app/src/main/java/com/codesoom/assignment/network/HttpMethod.java
@@ -1,0 +1,13 @@
+package com.codesoom.assignment.network;
+
+public enum HttpMethod {
+    GET, POST, PUT, PATCH, DELETE;
+
+    public static HttpMethod convert(String method) {
+        try {
+            return HttpMethod.valueOf(method);
+        } catch (final IllegalArgumentException e) {
+            return null;
+        }
+    }
+}
